* 'master' of https://bitbucket.org/lei-isep/switch-2017-g003:
  Added the parse of the ProjectState to the the LoadProjectData Class.
  Corrected the tests that were failing due to change of the type of the variables in the classes Task and ProjectCollabortor
  Changed the type of the parameter costPerEffort in the createProjectCollaborator method in the projectService to double
  Changed the type of the input parameter costPerEffort in the constructor of the ProjectCollaborator to double.
  Changed the type of the variables taskBudget and estimatedTaskEffort to double.

* 'master' of https://bitbucket.org/lei-isep/switch-2017-g003:
  Added the parse of the ProjectState to the the LoadProjectData Class.
  Corrected the tests that were failing due to change of the type of the variables in the classes Task and ProjectCollabortor
  Changed the type of the parameter costPerEffort in the createProjectCollaborator method in the projectService to double
  Changed the type of the input parameter costPerEffort in the constructor of the ProjectCollaborator to double.
  Changed the type of the variables taskBudget and estimatedTaskEffort to double.

+++ src/main/java/project/model/ProjectCollaborator.java
@@ -36,7 +36,7 @@ public class ProjectCollaborator implements Serializable {
 	 * @param costPerEffort
 	 *            how much the Project Collaborator costs per unit of effort
 	 */
-	public ProjectCollaborator(User collab, int costPerEffort) {
+	public ProjectCollaborator(User collab, double costPerEffort) {
 		this.collaborator = collab;
 		this.status = true;
 		this.costPerEffort = costPerEffort;
@@ -66,7 +66,7 @@ public class ProjectCollaborator implements Serializable {
 		return costPerEffort;
 	}
 
-	public void setCostPerEffort(int costPerEffort) {
+	public void setCostPerEffort(double costPerEffort) {
 		this.costPerEffort = costPerEffort;
 	}
